requirements.txt
requests ==2.25.1
requests-http-signature ==0.2.0
logzero ==1.6.3
pycryptodome ==3.10.1
cryptography ==3.4.6
black ==20.8b1
